@@ -1,16 +1,38 @@
 module github.com/popura-network/Popura
 
-go 1.16
+go 1.17
 
 require (
+	github.com/Arceliar/ironwood v0.0.0-20220409035209-b7f71f05435a
+	github.com/Arceliar/phony v0.0.0-20210209235338-dde1a8dca979
 	github.com/cheggaaa/pb/v3 v3.0.8
 	github.com/gologme/log v1.2.0
 	github.com/hashicorp/go-syslog v1.0.0
 	github.com/hjson/hjson-go v3.1.0+incompatible
 	github.com/kardianos/minwinsvc v1.0.0
-	github.com/miekg/dns v1.1.41 // indirect
 	github.com/mitchellh/mapstructure v1.4.1
-	github.com/yggdrasil-network/yggdrasil-go v0.4.3
+	github.com/vishvananda/netlink v1.1.0
+	github.com/yggdrasil-network/yggdrasil-go v0.4.4
 	github.com/zhoreeq/meshname v0.2.0
+	golang.org/x/mobile v0.0.0-20220112015953-858099ff7816
+	golang.org/x/net v0.0.0-20211101193420-4a448f8816b3
+	golang.org/x/sys v0.0.0-20211102192858-4dd72447c267
 	golang.org/x/text v0.3.8-0.20211004125949-5bd84dd9b33b
+	golang.zx2c4.com/wireguard v0.0.0-20211017052713-f87e87af0d9a
+	golang.zx2c4.com/wireguard/windows v0.4.12
+)
+
+require (
+	github.com/VividCortex/ewma v1.2.0 // indirect
+	github.com/fatih/color v1.12.0 // indirect
+	github.com/mattn/go-colorable v0.1.8 // indirect
+	github.com/mattn/go-isatty v0.0.13 // indirect
+	github.com/mattn/go-runewidth v0.0.13 // indirect
+	github.com/miekg/dns v1.1.41 // indirect
+	github.com/rivo/uniseg v0.2.0 // indirect
+	github.com/vishvananda/netns v0.0.0-20210104183010-2eb08e3e575f // indirect
+	golang.org/x/crypto v0.0.0-20210921155107-089bfa567519 // indirect
+	golang.org/x/mod v0.4.2 // indirect
+	golang.org/x/tools v0.1.8-0.20211022200916-316ba0b74098 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
 )
